Create pyc_utils.py with Python 2 compatible pyc functions
